+++ logging/test/unit/liberator/mixin/logging/core_test.clj
@@ -1,6 +1,8 @@
 (ns liberator.mixin.logging.core-test
   (:require
-   [clojure.test :refer [deftest is testing]]
+   [clojure.test :refer [deftest is]]
+
+   [cartus.test :as ct]
 
    [ring.mock.request :as ring]
 
@@ -9,27 +11,22 @@
    [liberator.mixin.core :as core]
    [liberator.mixin.logging.core :as logging]))
 
-(defrecord NullLogger []
-           logging/Logger
-           (log-error [_self _message _context _cause]))
-
 (defn call-resource [resource request]
   (->
     (resource request)
     (update :body jason-conv/<-wire-json)))
 
-(deftest with-logger
-  (testing "adds logger from dependencies to context"
-    (let [logger (->NullLogger)
-          resource (core/build-resource
-                     {:available-media-types ["application/json"]}
-                     (logging/with-logger {:logger logger})
-                     {:handle-ok
-                      (fn [{:keys [resource]}]
-                        (if (= ((get resource :logger)) logger)
-                          {:result true}
-                          {:result false}))})
-          response (call-resource
-                     resource
-                     (ring/request :get "/"))]
-      (is (= {:result true} (:body response))))))
+(deftest with-logger-adds-logger-from-dependencies-to-resource
+  (let [logger (ct/logger)
+        resource (core/build-resource
+                   {:available-media-types ["application/json"]}
+                   (logging/with-logger {:logger logger})
+                   {:handle-ok
+                    (fn [{:keys [resource]}]
+                      (if (= ((get resource :logger)) logger)
+                        {:result true}
+                        {:result false}))})
+        response (call-resource
+                   resource
+                   (ring/request :get "/"))]
+    (is (= {:result true} (:body response)))))